removed some delteed msgs, lots of other stuff
