+++ .pylintrc
@@ -1,4 +1,4 @@
 [MESSAGES CONTROL]
-disable=line-too-long, missing-function-docstring, missing-module-docstring, broad-except, anomalous-backslash-in-string, logging-fstring-interpolation, bare-except, protected-access, import-error
+disable=line-too-long, missing-function-docstring, missing-module-docstring, missing-class-docstring, broad-except, anomalous-backslash-in-string, logging-fstring-interpolation, bare-except, protected-access, import-error, too-few-public-methods, too-many-return-statements, arguments-renamed
 generated-members=config.*
 good-names=e
