+++ src/cogs/reaction_commands.py
@@ -1,18 +1,15 @@
-import asyncio
 import logging
 
 import discord
 from discord.ext import commands
 from discord.utils import get
 
+from utils.database.db import DatabaseManager as db
 import cogs.helpers.views.action_views as action_views
+from cogs.helpers.views import command_views
 import cogs.helpers.actions as actions
 from utils.others import Others
-from cogs.helpers.views import command_views
 import config
-
-from utils.database.db import DatabaseManager as db
-from utils.background import ScrapeChallenges, UpdateHelpers
 
 log = logging.getLogger(__name__)
 
@@ -87,7 +84,6 @@ For **help** tickets:
             return
 
         await actions.Utility.add(ctx.channel, member)
-        await Others.delmsg(ctx)
 
     @commands.command(name="remove", aliases=["r", "rm"])
     @commands.has_role(config.ADMIN_ROLE)
@@ -107,7 +103,6 @@ For **help** tickets:
             return
 
         await actions.Utility.remove(ctx.channel, member)
-        await Others.delmsg(ctx)
 
     @commands.command(name="close", aliases=["cl"])
     async def close(self, ctx):
@@ -148,48 +143,6 @@ For **help** tickets:
 
         await Others.delmsg(ctx)
 
-    @commands.command(name="purge")
-    @commands.has_role(config.ADMIN_ROLE)
-    async def purge(self, ctx, limit: int):
-        """purges x amount of messages"""
-
-        await ctx.channel.purge(limit=limit + 1)
-        message = await ctx.send(f'Purged {limit} messages')
-        await asyncio.sleep(3)
-        await message.delete()
-
-    @commands.command(name="check")
-    @commands.has_role(config.ADMIN_ROLE)
-    async def check_discord(self, ctx):
-        """Checks if all configurations are valid"""
-        bot_guild = ctx.guild.get_member(self.bot.user.id)
-        checks = {"ticket ping role": bool(get(ctx.guild.roles, name=config.TICKET_PING_ROLE)),
-                  "bots role": bool(get(ctx.guild.roles, name=config.BOTS_ROLE)),
-                  "helpers role": bool(get(ctx.guild.roles, name=config.HELPER_ROLE)),
-                  "channel log category": bool(get(ctx.guild.categories, name=config.LOG_CHANNEL_CATEGORY)),
-                  "channel log name": bool(get(ctx.guild.text_channels, name=config.LOG_CHANNEL_NAME)),
-                  "is admin": bool(bot_guild.guild_permissions.administrator)}
-
-        def check_all(return_print=False):
-            if all(checks.values()):
-                return True
-
-            failures = [check for check, status
-                        in checks.items() if status is False]
-
-            if return_print:
-                return failures
-            return False
-
-        failure = check_all()
-        if not failure:
-            fails = "\n".join(check_all(return_print=True))
-            embed = Others.Embed(title="Failed Checks", description=fails)
-            await ctx.channel.send(embed=embed)
-            return
-
-        await ctx.channel.send("All checks were successful 😎")
-
     @commands.command(name="transcript", alias=["tsc"])
     @commands.has_role(config.ADMIN_ROLE)
     @commands.cooldown(rate=1, per=10, type=commands.BucketType.default)
@@ -228,91 +181,6 @@ For **help** tickets:
             title="Auto Message", description=f"{random_admin.mention} said the auto close message in {channel.mention}")
         embed.set_author(name=f"{ctx.author}",
                          icon_url=f"{ctx.author.avatar.url}")
-        await ctx.channel.send(embed=embed)
-        await Others.delmsg(ctx)
-
-    @commands.command(name="refresh", aliases=["ref"])
-    @commands.has_role(config.ADMIN_ROLE)
-    async def refresh(self, ctx):
-        """refreshes challenges from the api"""
-        embed = Others.Embed(
-            description="sending requests...")
-        message = await ctx.channel.send(embed=embed)
-
-        await ScrapeChallenges.main(self.bot)
-
-        embed.description = "challenges refreshed"
-        await message.edit(embed=embed)
-        await Others.delmsg(ctx)
-
-    @commands.group(name="helper", aliases=["h"], invoke_without_command=True)
-    @commands.has_role(config.ADMIN_ROLE)
-    async def helper(self, ctx):
-        """Base helper command. Shows stats on helpers."""
-        helper_role = get(ctx.guild.roles, name=config.HELPER_ROLE)
-        if len(helper_role.members):
-            helpers = '\n'.join(
-                [member.mention for member in helper_role.members])
-        else:
-            helpers = 'No helpers'
-
-        embed = Others.Embed(
-            title=f"{config.HELPER_ROLE}", description=helpers)
-        await ctx.channel.send(embed=embed)
-        await Others.delmsg(ctx)
-
-    @helper.command(name="add", aliases=["a"])
-    @commands.has_role(config.ADMIN_ROLE)
-    async def helper_add(self, ctx, member: discord.Member):
-        """adds a helper"""
-        helper_role = get(member.guild.roles, name=config.HELPER_ROLE)
-        helper_ids = [helper.id for helper in helper_role.members]
-        if member.id in helper_ids:
-            embed = Others.Embed(
-                description=f"Member {member.mention} already has role {config.HELPER_ROLE}")
-            await ctx.channel.send(embed=embed)
-            return
-
-        await member.add_roles(helper_role)
-        embed = Others.Embed(
-            description=f"Added {member.mention} to role {config.HELPER_ROLE}")
-        await ctx.channel.send(embed=embed)
-        await Others.delmsg(ctx)
-
-    @helper.command(name="remove", aliases=["r", "rm"])
-    @commands.has_role(config.ADMIN_ROLE)
-    async def helper_remove(self, ctx, member: discord.Member):
-        """removes a helper"""
-        helper_role = get(member.guild.roles, name=config.HELPER_ROLE)
-        helper_ids = [helper.id for helper in helper_role.members]
-        if member.id not in helper_ids:
-            embed = Others.Embed(
-                description=f"Member {member.mention} does not have role {config.HELPER_ROLE}")
-            await ctx.channel.send(embed=embed)
-            return
-        await member.remove_roles(helper_role)
-        embed = Others.Embed(
-            description=f"Removed {member.mention} from role {config.HELPER_ROLE}")
-        await ctx.channel.send(embed=embed)
-        await Others.delmsg(ctx)
-
-    @helper.command(name="refresh", aliases=["ref"])
-    @commands.has_role(config.ADMIN_ROLE)
-    async def helper_refresh(self, ctx):
-        """refreshes helpers from the api"""
-        await UpdateHelpers.main(self.bot)
-        embed = Others.Embed(
-            description="helpers refreshed")
-        await ctx.channel.send(embed=embed)
-        await Others.delmsg(ctx)
-
-    @helper.command(name="update", aliases=["upd"])
-    @commands.has_role(config.ADMIN_ROLE)
-    async def helper_update(self, ctx):
-        """updates helpers to channels"""
-        await UpdateHelpers.add_helpers(self.bot)
-        embed = Others.Embed(
-            description="helpers updated")
         await ctx.channel.send(embed=embed)
         await Others.delmsg(ctx)
